+++ apps/prairielearn/src/tests/fileEditor.test.js
@@ -617,9 +617,7 @@ function verifyEdit(
     assert.isString(locals.file_edit_orig_hash);
   });
   it('should have a script with draft file contents', function (callback) {
-    elemList = locals.$('script');
-    for (let i = 0; i < elemList.length; i++) {
-      let elem = elemList[i];
+    for (const elem of Array.from(locals.$('script'))) {
       if (typeof elem !== 'undefined' && Object.prototype.hasOwnProperty.call(elem, 'children')) {
         if (elem.children.length > 0) {
           if (Object.prototype.hasOwnProperty.call(elem.children[0], 'data')) {
@@ -648,9 +646,7 @@ function verifyEdit(
     }
   });
   it(`should have a script with disk file contents - ${expectedToFindChoice}`, function (callback) {
-    elemList = locals.$('script');
-    for (let i = 0; i < elemList.length; i++) {
-      let elem = elemList[i];
+    for (const elem of Array.from(locals.$('script'))) {
       if (typeof elem !== 'undefined' && Object.prototype.hasOwnProperty.call(elem, 'children')) {
         if (elem.children.length > 0) {
           if (Object.prototype.hasOwnProperty.call(elem.children[0], 'data')) {