+++ packages/csv/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/csv
 
+## 1.0.3
+
+### Patch Changes
+
+- 098f581da: Upgrade all dependencies
+
+## 1.0.2
+
+### Patch Changes
+
+- 2b003b4d9: Upgrade all dependencies
+
 ## 1.0.1
 
 ### Patch Changes