+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/postgres-tools
 
+## 1.1.4
+
+### Patch Changes
+
+- 098f581da: Upgrade all dependencies
+- Updated dependencies [098f581da]
+  - @prairielearn/postgres@1.7.3
+
+## 1.1.3
+
+### Patch Changes
+
+- 2b003b4d9: Upgrade all dependencies
+- Updated dependencies [2b003b4d9]
+  - @prairielearn/postgres@1.7.2
+
 ## 1.1.2
 
 ### Patch Changes