+++ packages/sentry/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/sentry
 
+## 1.0.6
+
+### Patch Changes
+
+- 098f581da: Upgrade all dependencies
+
+## 1.0.5
+
+### Patch Changes
+
+- 2b003b4d9: Upgrade all dependencies
+
 ## 1.0.4
 
 ### Patch Changes